fix index.ts and dropdown widget
